requirements.txt
streamlit>=1.22.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.20
albumentations>=1.3.0
opencv-python-headless>=4.7.0.72
numpy>=1.21.0
pillow>=9.0.0
